fix: extra safety net to block .zsh_history
Sometimes, the zsh_history was committed, likely due to errors in the
zshrc, blocking the setting of the history location outside the public
dotfiles repo. None included anything remotely critical, but just to be
safe, ran git-filter-repo on it regardless, and added extra .git_ignores
to block any zsh_history file.